# agentloom-server/sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/platform-api-tokens": {
      "post": {
        "operationId": "PlatformApiToken_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformApiTokenSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token 创建成功，仅此次返回明文",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformApiTokenCreateEnvelopeSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          },
          "403": {
            "description": "权限不足"
          },
          "409": {
            "description": "Token 数量超限"
          },
          "422": {
            "description": "请求参数校验失败"
          }
        },
        "security": [
          {
            "X-Api-Key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "创建 API Token",
        "tags": [
          "Platform API Tokens"
        ]
      },
      "get": {
        "operationId": "PlatformApiToken_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "type": "string",
              "enum": [
                "active",
                "revoked",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token 列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformApiTokenListResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "X-Api-Key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "分页查询当前用户的 API Token 列表",
        "tags": [
          "Platform API Tokens"
        ]
      }
    },
    "/api/v1/platform-api-tokens/{tokenId}": {
      "delete": {
        "operationId": "PlatformApiToken_revoke",
        "parameters": [
          {
            "name": "tokenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Token 已撤销"
          },
          "401": {
            "description": "认证失败"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "Token 不存在"
          },
          "409": {
            "description": "Token 已被撤销"
          }
        },
        "security": [
          {
            "X-Api-Key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "撤销 API Token",
        "tags": [
          "Platform API Tokens"
        ]
      }
    },
...
  },
  "info": {
    "title": "AgentLoom API",
    "description": "AgentLoom 多智能体协作平台 API",
    "version": "1.0",
    "contact": {
      "name": "AgentLoom",
      "url": "https://agentloom.dev",
      "email": "support@agentloom.dev"
    },
    "license": {
      "name": "Proprietary"
    }
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "X-Api-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
...
      "CreatePlatformApiTokenSwaggerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "scopes": {
            "type": "string",
            "maxLength": 1024
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "name"
        ]
      },
      "PlatformApiTokenCreateEnvelopeSwaggerDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              },
              "tokenPrefix": {
                "type": "string"
              },
              "scopes": {
                "type": "string",
                "nullable": true
              },
              "lastUsedAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "nullable": true
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "nullable": true
              },
              "isRevoked": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "token": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "tokenPrefix",
              "scopes",
              "lastUsedAt",
              "expiresAt",
              "isRevoked",
              "createdAt",
              "token"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "PlatformApiTokenListResponseSwaggerDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "name": {
                  "type": "string"
                },
                "tokenPrefix": {
                  "type": "string"
                },
                "scopes": {
                  "type": "string",
                  "nullable": true
                },
                "lastUsedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "nullable": true
                },
                "expiresAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "nullable": true
                },
                "isRevoked": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "name",
                "tokenPrefix",
                "scopes",
                "lastUsedAt",
                "expiresAt",
                "isRevoked",
                "createdAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "pageSize": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "total": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "page",
              "pageSize",
              "total"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      }
    }
  }
}